requirements.txt
streamlit==1.28.1
requests==2.31.0
pandas==2.0.3
matplotlib==3.8.1
reportlab==4.0.7
fastapi==0.104.1
uvicorn==0.24.0
joblib==1.3.2
scikit-learn==1.3.2
shap==0.42.3
numpy==1.24.3
openai==1.3.5
python-dotenv==1.0.0
langchain==0.0.350
